refactor(osb-core): extracted all security configurations to a separate package to be more flexible in reuse scenarios of other projects (autoscaler, log-metric, backup-manager). Refactored model directory in regard of catalog and plan packages
