+++ bosh/src/main/java/de/evoila/cf/cpi/bosh/deployment/DeploymentManager.java
@@ -7,6 +7,10 @@ import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import de.evoila.cf.broker.bean.BoshProperties;
 import de.evoila.cf.broker.model.*;
+import de.evoila.cf.broker.model.catalog.plan.InstanceGroupConfig;
+import de.evoila.cf.broker.model.catalog.plan.Metadata;
+import de.evoila.cf.broker.model.catalog.plan.NetworkReference;
+import de.evoila.cf.broker.model.catalog.plan.Plan;
 import de.evoila.cf.broker.util.GlobalConstants;
 import de.evoila.cf.broker.util.MapUtils;
 import de.evoila.cf.cpi.bosh.deployment.manifest.InstanceGroup;
