+++ bosh/src/main/java/de/evoila/cf/cpi/bosh/deployment/manifest/instanceGroup/InstanceJob.java
@@ -1,7 +1,7 @@
 package de.evoila.cf.cpi.bosh.deployment.manifest.instanceGroup;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import de.evoila.cf.broker.model.NetworkReference;
+import de.evoila.cf.broker.model.catalog.plan.NetworkReference;
 import de.evoila.cf.broker.model.VolumeUnit;
 import de.evoila.cf.cpi.bosh.deployment.manifest.Update;
