+++ existing-service/src/main/java/de/evoila/cf/broker/bean/ExistingEndpointBean.java
@@ -1,6 +1,6 @@
 package de.evoila.cf.broker.bean;
 
-import de.evoila.cf.broker.model.ServerAddress;
+import de.evoila.cf.broker.model.catalog.ServerAddress;
 
 import java.util.List;
